# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Reloop API
  version: 1.0.0
  description: |
    API oficial del backend de Reloop.
    Incluye autenticación JWT, publicaciones, ofertas, carrito,
    compras, favoritos, ratings, usuarios y comentarios.
servers:
  - url: http://localhost:3000
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  #############################
  # GLOBAL SCHEMAS
  #############################
  schemas:
...
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string, example: vale@example.com }
        username: { type: string }
        name: { type: string, nullable: true }
        lastName: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        gender: { type: string, nullable: true }
        country: { type: string, nullable: true }
        birthDate:
          type: string
          format: date
          nullable: true
        avatar:
          type: string
          example: https://example.com/avatar.png
        createdAt:
          type: string
          format: date-time
...
    ListingPhoto:
      type: object
      properties:
        id: { type: string }
        url: { type: string }
    Listing:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
        price: { type: number }
        condition: { type: string }
        category: { type: string }
        subCategory: { type: string, nullable: true }
        subSubCategory: { type: string, nullable: true }
        brand: { type: string, nullable: true }
        color: { type: string, nullable: true }
        seller:
          $ref: "#/components/schemas/User"
        photos:
          type: array
          items:
            $ref: "#/components/schemas/ListingPhoto"
        createdAt:
          type: string
          format: date-time
...
    Offer:
      type: object
      properties:
        id: { type: string }
        amount: { type: number }
        status:
          type: string
          enum: [PENDING, ACCEPTED, REJECTED, COUNTERED]
        counterOfferAmount:
          type: number
          nullable: true
        listing:
          $ref: "#/components/schemas/Listing"
        buyer:
          $ref: "#/components/schemas/User"
        seller:
          $ref: "#/components/schemas/User"
...
    OrderItem:
      type: object
      properties:
        id: { type: string }
        price: { type: number }
        listing:
          $ref: "#/components/schemas/Listing"
    Order:
      type: object
      properties:
        id: { type: string }
        totalAmount: { type: number }
        createdAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
...
paths:
...
  /offers/{listingId}:
    post:
      tags: [Offers]
      summary: Crear oferta sobre un listing
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: listingId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount:
                  type: number
                  example: 1500
      responses:
        "201":
          description: Oferta creada correctamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
        "404":
          description: listing_not_found
        "400":
          description: cannot_offer_on_own_listing
        "401":
          description: Token inválido o faltante
  /offers/mine:
    get:
      tags: [Offers]
      summary: Ver ofertas hechas por el usuario autenticado
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista de ofertas enviadas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Offer"
  /offers/received:
    get:
      tags: [Offers]
      summary: Ver ofertas recibidas en mis publicaciones
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista de ofertas recibidas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Offer"
  /offers/respond/{offerId}:
    patch:
      tags: [Offers]
      summary: Responder una oferta (aceptar, rechazar o contraofertar)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: offerId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [ACCEPTED, REJECTED, COUNTERED]
                counterAmount:
                  type: number
                  nullable: true
                  example: 1800
      responses:
        "200":
          description: Oferta actualizada correctamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
        "404":
          description: offer_not_found
        "403":
          description: not_your_offer
        "401":
          description: Token inválido o faltante
  ###############################################
  # ORDERS
  ###############################################
  /orders:
    post:
      tags: [Orders]
      summary: Crear orden de compra (carrito o compra directa)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [listingIds]
              properties:
                listingIds:
                  type: array
                  items:
                    type: string
                  example: ["cl123", "cl456"]
      responses:
        "201":
          description: Orden creada correctamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: invalid_listing_ids
        "401":
          description: Token inválido o faltante

    get:
      tags: [Orders]
      summary: Listar compras del usuario autenticado
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista de órdenes realizadas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          description: Token inválido o faltante
  ###############################################
  # RATINGS
  ###############################################
...